Move all source files to under eviden/ directory
