+++ eviden/core.py
